+++ README.md
@@ -13,3 +13,11 @@ vue: {
   parent: components/sign_up/SignUp.vue
   child: components/sign_up/SignUpContens.vue
 }
+
+# パスワード再設定ページ
+rails: users/passwords#new
+js: javascript/packs/reset_password.js
+vue: {
+  parent: components/reset_password/ResetPassword.vue
+  child: components/reset_password/ResetPasswordContents.vue
+}
