+++ app/javascript/components/reset_password/ResetPassword.vue
@@ -1,42 +1,38 @@
 <template>
-  <div id="app">
-    <Header></Header>
-    <div class="main-box">
+  <main id="app" class="container">
+    <div class="main-conts">
       <ResetPasswordContents></ResetPasswordContents>
     </div>
-  </div>
+  </main>
 </template>
 
 <script>
-import Header from 'components/Header.vue';
 import ResetPasswordContents from 'components/reset_password/ResetPasswordContents.vue';
-import 'flower.png'
 
 export default {
   components: {
-    Header, ResetPasswordContents
+    ResetPasswordContents
   },
   data: function () {
     return {
-      message: "Hello Vue!"
     }
   }
 }
 </script>
 
 <style scoped>
-#app {
-  background-image: url('~flower.png');
-  background-color:rgba(255,255,255,0.8);
-  background-blend-mode:lighten;
+.container {
+  height: 100vh;
+  background: linear-gradient(134.5deg, #F9516F 11.31%, #FF8F6B 87.66%);
+
 }
-.main-box {
+
+.main-conts{
+  height: 100vh;
   display: flex;
-  height: calc( 100vh - 100px);
-  width: 100%;
-}
-p {
-  font-size: 2em;
-  text-align: center;
+  justify-content: center;
+  align-items: center;
+  font-size: 14px;
+  font-family: Noto Sans CJK JP;
 }
 </style>
